Adds video styles
Creates `video` class which you place on the section wrapper. This
tells the section and column to be 370px high and 100% wide.
The `video` class also tells the section and column to not have any
margins and to be position absolute within the wrapper
The p and a tag are absolute with 100% width and 370px high
The a tag has a play icon and is centered in the middle of the section
with a default state and a hover state image